feat: allow customising upload url and upload id extraction
