+++ packages/server/src/handlers/BaseHandler.ts
@@ -1,16 +1,14 @@
 import EventEmitter from 'node:events'
 
 import type {ServerOptions} from '../types'
-import type {DataStore} from '../models'
+import type {DataStore, UploadIdGenerator} from '../models'
 import type http from 'node:http'
-
-const reExtractFileID = /([^/]+)\/?$/
-const reForwardedHost = /host="?([^";]+)/
-const reForwardedProto = /proto=(https?)/
+import {DefaultUploadIdGenerator} from '../models'
 
 export class BaseHandler extends EventEmitter {
   options: ServerOptions
   store: DataStore
+  uploadIdGenerator: UploadIdGenerator
 
   constructor(store: DataStore, options: ServerOptions) {
     super()
@@ -20,6 +18,13 @@ export class BaseHandler extends EventEmitter {
 
     this.store = store
     this.options = options
+    this.uploadIdGenerator =
+      options.uploadIdGenerator ??
+      new DefaultUploadIdGenerator({
+        path: options.path,
+        relativeLocation: options.relativeLocation,
+        respectForwardedHeaders: options.respectForwardedHeaders,
+      })
   }
 
   write(res: http.ServerResponse, status: number, headers = {}, body = '') {
@@ -33,49 +38,10 @@ export class BaseHandler extends EventEmitter {
   }
 
   generateUrl(req: http.IncomingMessage, id: string) {
-    id = encodeURIComponent(id)
-
-    const forwarded = req.headers.forwarded as string | undefined
-    const path = this.options.path === '/' ? '' : this.options.path
-    // @ts-expect-error baseUrl type doesn't exist?
-    const baseUrl = req.baseUrl ?? ''
-    let proto
-    let host
-
-    if (this.options.relativeLocation) {
-      return `${baseUrl}${path}/${id}`
-    }
-
-    if (this.options.respectForwardedHeaders) {
-      if (forwarded) {
-        host ??= reForwardedHost.exec(forwarded)?.[1]
-        proto ??= reForwardedProto.exec(forwarded)?.[1]
-      }
-
-      const forwardHost = req.headers['x-forwarded-host']
-      const forwardProto = req.headers['x-forwarded-proto']
-
-      // @ts-expect-error we can pass undefined
-      if (['http', 'https'].includes(forwardProto)) {
-        proto ??= forwardProto as string
-      }
-
-      host ??= forwardHost
-    }
-
-    host ??= req.headers.host
-    proto ??= 'http'
-
-    return `${proto}://${host}${baseUrl}${path}/${id}`
+    return this.uploadIdGenerator.generateUrl(req, id)
   }
 
   getFileIdFromRequest(req: http.IncomingMessage) {
-    const match = reExtractFileID.exec(req.url as string)
-
-    if (!match || this.options.path.includes(match[1])) {
-      return false
-    }
-
-    return decodeURIComponent(match[1])
+    return this.uploadIdGenerator.getFileIdFromRequest(req)
   }
 }
