+++ packages/server/src/models/index.ts
@@ -3,3 +3,8 @@ export * as Metadata from './Metadata'
 export {StreamSplitter} from './StreamSplitter'
 export {Uid} from './Uid'
 export {Upload} from './Upload'
+export {
+  DefaultUploadIdGeneratorOptions,
+  DefaultUploadIdGenerator,
+  UploadIdGenerator,
+} from './UploadIdGenerator'
